[feature] Refactor no programmed shares and inspections (#7)
* Refactor no programmed shares and interventions

* Refactor inspections

* Refactor inspections

* Close first phase

+++ src/main/webapp/ui/jsp/view/shares/intervention.jsp
@@ -204,10 +204,6 @@ em[disabled] {
 
 <input id="userLanguageInput" value="${ language }" type="hidden" />
 
-<form id="createNoProgForm">
-	<input id="noPrProjectId" name="projectId" type="hidden" required>
-</form>
-
 <form id="createShareForm">
 	<input id="shareProjectId" name="projectId" type="hidden" required>
 	<input id="shareClientNotif" name="clientNotif" type="hidden" required>
@@ -253,7 +249,7 @@ em[disabled] {
 				<div class="modal-footer clearfix">
 					<div class="w-100">
 						<div class="float-left">
-							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
+							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="close" /></button>
 						</div>
 					</div>
 				</div>
@@ -284,42 +280,12 @@ em[disabled] {
 			<div class="modal-footer clearfix">
 				<div class="w-100">
 					<div class="float-left">
-						<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
+						<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="close" /></button>
 					</div>
 					
 					<div class="float-right">
 						<button id="btnRefuseNotif" type="button" class="btn btn-sm btn-danger"><spring:message code="no" /></button>
 						<button id="btnConfirmNotif" type="button" class="btn btn-sm btn-success"><spring:message code="yes" /></button>
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
-</div>
-<!-- /MODAL -->
-
-<!-- MODAL -->
-<div id="popupModal" class="modal fade" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
-	<div class="modal-dialog modal-lg">
-		<div class="modal-content">
-			<div class="modal-header">
-				<div class="modal-title">
-					<h5 id="modalTitle">
-						<spring:message code="info" />
-					</h5>
-				</div>
-			</div>
-			<div class="modal-body">		
-				<div class="row">
-					<div class="col">
-						<span><spring:message code="shares.intervention.create.not.station" /></span>
-					</div>
-				</div>
-			</div>
-			<div class="modal-footer clearfix">
-				<div class="w-100">
-					<div class="float-left">
-						<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
 					</div>
 				</div>
 			</div>
@@ -404,7 +370,7 @@ em[disabled] {
 				<div class="modal-footer clearfix">
 					<div class="w-100">
 						<div class="float-left">
-							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
+							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="close" /></button>
 						</div>
 						<div class="float-right">
 							<button id="createConstBtn" type="button" class="btn btn-sm btn-success"><spring:message code="finish" /></button>
@@ -519,7 +485,7 @@ em[disabled] {
 				<div class="modal-footer clearfix">
 					<div class="w-100">
 						<div class="float-left">
-							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
+							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="close" /></button>
 						</div>
 						<div class="float-right">
 							<button id="createNoPrShareBtn" type="button" class="btn btn-sm btn-info"><spring:message code="shares.intervention.create.nopr.title" /></button>
@@ -609,7 +575,7 @@ em[disabled] {
 				<div class="modal-footer clearfix">
 					<div class="w-100">
 						<div class="float-left">
-							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="cerrar" /></button>
+							<button type="button" class="btn btn-sm" data-dismiss="modal"><spring:message code="close" /></button>
 						</div>
 						<div class="float-right">
 							<button id="createWorkBtn" type="button" class="btn btn-sm btn-success"><spring:message code="finish" /></button>
@@ -851,35 +817,9 @@ em[disabled] {
 		});
 
 		$('#btnNoProgrammedShare').click(function () {
-
-			var dropdownInfo = $('#projectDropdown').find(':selected').attr('data-info');
-
-			var station = dropdownInfo.split(';')[0];
-
-			if (station != 1) {
-				$('#popupModal').modal('show');
-			} else {
-
-				showLoading();
-				
-				var projectId = $('#projectDropdown').val();
-				
-				$('#noPrProjectId').val(projectId);
-	
-				$.ajax({
-					type: "POST",
-					url: "/shares/intervention/no-programmed/create",
-					data: $('#createNoProgForm').serialize(),
-					success: function(msg) {
-						hideLoading();
-						window.location.href = "/shares/intervention/no-programmed/detail/" + msg;
-					},
-					error: function(e) {
-						hideLoading();
-						showNotify(e.responseText, 'danger');
-					}
-				});
-			}
+			const userId = ${ user.id };
+			const projectId = $('#projectDropdown').val();
+			createNoProgrammedShare(userId, projectId);
 
 			$('#createSelectorModal').modal('hide');
 		});
@@ -888,7 +828,6 @@ em[disabled] {
 			shareSelectedType = 'ws';
 			createWorkShare();
 			$('#createSelectorModal').modal('hide');
-			
 		});
 		/* END INITIALIZE SHARE */
 
@@ -1037,24 +976,9 @@ em[disabled] {
 		        cache: false,
 				data: data,
 				success: function(msg) {
-
-					var projectId = $('#projectIdProgInput').val();
-					
-					$('#noPrProjectId').val(projectId);
-
-					$.ajax({
-						type: "POST",
-						url: "/shares/intervention/no-programmed/create",
-						data: $('#createNoProgForm').serialize(),
-						success: function(msg) {
-							window.location.href = "/shares/intervention/no-programmed/detail/" + msg;
-						},
-						error: function(e) {
-							hideLoading();
-							showNotify(e.responseText, 'danger');
-						}
-					});
-					
+					const userId = ${ user.id };
+					const projectId = $('#projectIdProgInput').val();
+					createNoProgrammedShare(userId, projectId);
 				},
 				error: function(e) {
 					hideLoading();
@@ -1347,24 +1271,30 @@ em[disabled] {
 			if (shareType === 'cs') {
 				restUrl = '/shares/intervention/construction/delete/' + shareId;
 			} else if (shareType === 'is') {
-				restUrl = '/shares/intervention/no-programmed/delete/' + shareId;
+				axios.delete('/v1/shares/no-programmed/' + shareId).then(() => {
+					$('#dTable').DataTable().ajax.reload();
+					showNotify(messages.shares.noprogrammed.delete.success.replace('{0}', shareId));
+				}).catch(error => showNotify(error, 'danger'))
+						.finally(() => hideLoading());
 			} else if (shareType === 'ips') {
 				restUrl = '/shares/intervention/programmed/delete/' + shareId;
 			} else if (shareType === 'ws') {
 				restUrl = '/shares/work/' + shareId;
 			}			
 
-			$.ajax({
-				type: "DELETE",
-				url: restUrl,
-				success: function(msg) {
-					$('#dTable').DataTable().ajax.reload();
-					showNotify(msg, 'success');
-				},
-				error: function(e) {
-					showNotify(e.responseText, 'danger');
-				}
-			});
+			if (restUrl) {
+				$.ajax({
+					type: "DELETE",
+					url: restUrl,
+					success: function (msg) {
+						$('#dTable').DataTable().ajax.reload();
+						showNotify(msg, 'success');
+					},
+					error: function (e) {
+						showNotify(e.responseText, 'danger');
+					}
+				});
+			}
 		}
 	}
